pyproject.toml
[tool.poetry]
name = "tt-reservations"
version = "0.1.0"
description = "This project wraps the mytischtennis.de booking page to make it easier to reserve playing times"
authors = ["Kory291 <01lukasschaefer@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.46.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"

[tool.poetry.scripts]
run-main = "tt_reservations.main:main"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
